# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AtsCheckRequest {
  /** The full resume text content */
  resumeText: string;
  /** Optional job description to match against. If omitted, a general ATS check is performed. */
  jobDescription?: string;
}

export interface AtsCheckResult {
  /** Overall ATS compatibility score (0-100) */
  overallScore: number;
  /** Keyword match score (0-100) */
  keywordScore: number;
  /** Format/structure score (0-100) */
  formatScore: number;
  /** Experience relevance score (0-100) */
  experienceScore: number;
  /** Keywords found in both resume and job description */
  matchedKeywords: string[];
  /** Important keywords from job description missing in resume */
  missingKeywords: string[];
  /** Actionable improvement suggestions */
  suggestions: string[];
  /** Strong points of the resume for this job */
  strengths: string[];
}

export interface ResumeFixRequest {
  /** The full plain-text resume content */
  resumeText: string;
  /** Optional job description used in the ATS check */
  jobDescription?: string;
  /** Keywords to incorporate from the ATS analysis */
  missingKeywords?: string[];
  /** ATS suggestions to address */
  suggestions?: string[];
}

export interface ExperienceImprovement {
  /** Index of the experience entry (0-based) */
  index: number;
  /** Improved bullet points for this experience entry */
  improvedBullets: string[];
}

export interface ResumeFixResult {
  /** AI-improved professional summary */
  improvedSummary: string;
  /** AI-improved skills section with missing keywords incorporated */
  improvedSkills: string;
  /** Improved bullet points per experience entry */
  experienceImprovements: ExperienceImprovement[];
  /** Brief explanation of changes made */
  overallChanges: string;
}

export interface ResumeTemplate {
  id: string;
  name: string;
  description: string;
}

export interface ResumeTemplateList {
  templates: ResumeTemplate[];
}