# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: resume
    description: Resume and ATS operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /resume/ats-check:
    post:
      operationId: atsCheck
      tags: [resume]
      summary: Run ATS analysis
      description: Analyzes a resume against a job description for ATS compatibility
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AtsCheckRequest"
      responses:
        "200":
          description: ATS analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AtsCheckResult"

  /resume/fix:
    post:
      operationId: fixResume
      tags: [resume]
      summary: Fix resume using AI based on ATS results
      description: Uses AI to rewrite and improve resume sections based on ATS analysis findings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResumeFixRequest"
      responses:
        "200":
          description: AI-improved resume sections
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeFixResult"

  /resume/templates:
    get:
      operationId: getResumeTemplates
      tags: [resume]
      summary: Get available resume templates
      responses:
        "200":
          description: List of templates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeTemplateList"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    AtsCheckRequest:
      type: object
      properties:
        resumeText:
          type: string
          description: The full resume text content
        jobDescription:
          type: string
          description: Optional job description to match against. If omitted, a general ATS check is performed.
      required:
        - resumeText

    AtsCheckResult:
      type: object
      properties:
        overallScore:
          type: number
          description: Overall ATS compatibility score (0-100)
        keywordScore:
          type: number
          description: Keyword match score (0-100)
        formatScore:
          type: number
          description: Format/structure score (0-100)
        experienceScore:
          type: number
          description: Experience relevance score (0-100)
        matchedKeywords:
          type: array
          items:
            type: string
          description: Keywords found in both resume and job description
        missingKeywords:
          type: array
          items:
            type: string
          description: Important keywords from job description missing in resume
        suggestions:
          type: array
          items:
            type: string
          description: Actionable improvement suggestions
        strengths:
          type: array
          items:
            type: string
          description: Strong points of the resume for this job
      required:
        - overallScore
        - keywordScore
        - formatScore
        - experienceScore
        - matchedKeywords
        - missingKeywords
        - suggestions
        - strengths

    ResumeFixRequest:
      type: object
      properties:
        resumeText:
          type: string
          description: The full plain-text resume content
        jobDescription:
          type: string
          description: Optional job description used in the ATS check
        missingKeywords:
          type: array
          items:
            type: string
          description: Keywords to incorporate from the ATS analysis
        suggestions:
          type: array
          items:
            type: string
          description: ATS suggestions to address
      required:
        - resumeText

    ResumeFixResult:
      type: object
      properties:
        improvedSummary:
          type: string
          description: AI-improved professional summary
        improvedSkills:
          type: string
          description: AI-improved skills section with missing keywords incorporated
        experienceImprovements:
          type: array
          items:
            $ref: "#/components/schemas/ExperienceImprovement"
          description: Improved bullet points per experience entry
        overallChanges:
          type: string
          description: Brief explanation of changes made
      required:
        - improvedSummary
        - improvedSkills
        - experienceImprovements
        - overallChanges

    ExperienceImprovement:
      type: object
      properties:
        index:
          type: integer
          description: Index of the experience entry (0-based)
        improvedBullets:
          type: array
          items:
            type: string
          description: Improved bullet points for this experience entry
      required:
        - index
        - improvedBullets

    ResumeTemplate:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
      required:
        - id
        - name
        - description

    ResumeTemplateList:
      type: object
      properties:
        templates:
          type: array
          items:
            $ref: "#/components/schemas/ResumeTemplate"
      required:
        - templates