# lib/api-zod/src/generated/types/experienceImprovement.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ExperienceImprovement {
  /** Index of the experience entry (0-based) */
  index: number;
  /** Improved bullet points for this experience entry */
  improvedBullets: string[];
}