# lib/api-zod/src/generated/types/resumeFixRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ResumeFixRequest {
  /** The full plain-text resume content */
  resumeText: string;
  /** Optional job description used in the ATS check */
  jobDescription?: string;
  /** Keywords to incorporate from the ATS analysis */
  missingKeywords?: string[];
  /** ATS suggestions to address */
  suggestions?: string[];
}