# lib/api-zod/src/generated/types/resumeFixResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ExperienceImprovement } from "./experienceImprovement";

export interface ResumeFixResult {
  /** AI-improved professional summary */
  improvedSummary: string;
  /** AI-improved skills section with missing keywords incorporated */
  improvedSkills: string;
  /** Improved bullet points per experience entry */
  experienceImprovements: ExperienceImprovement[];
  /** Brief explanation of changes made */
  overallChanges: string;
}